ore: Introspect disk for size rather than hardcode 8GB
We want to expand the disk size of Red Hat CoreOS.  Use the JSON
output of `qemu-img info` to get the size of the VMDK, and pass
that on to AWS.

The hardcoded 8GB moves to Plume.  I think to support larger disks
there then we'd need to query the snapshot or so?